Move numismatics from AdvancedController to CatalogController
Newer versions of Blacklight Advanced Search don't include an
AdvancedController, so this helps us move away from it.

Taken from c5663fb and 24b3fc6
