+++ app/views/advanced/_numismatics_advanced_search_form.html.erb
@@ -11,21 +11,21 @@
         <%# Indicate that this a numismatics advanced search %>
         <%= render Blacklight::HiddenSearchStateComponent.new(params: {'advanced_type' => 'numismatics'}) %>
         <div>
-          <%= render partial: 'facet_as_select', locals: { facet_field_name: 'issue_object_type_s' } %>
-          <%= render partial: 'facet_as_select', locals: { facet_field_name: 'issue_denomination_s' } %>
-          <%= render partial: 'facet_as_select', locals: { facet_field_name: 'issue_metal_s' } %>
-          <%= render partial: 'facet_as_select', locals: { facet_field_name: 'issue_city_s' } %>
-          <%= render partial: 'facet_as_select', locals: { facet_field_name: 'issue_state_s' } %>
-          <%= render partial: 'facet_as_select', locals: { facet_field_name: 'issue_region_s' } %>
+          <%= render partial: 'advanced/facet_as_select', locals: { facet_field_name: 'issue_object_type_s' } %>
+          <%= render partial: 'advanced/facet_as_select', locals: { facet_field_name: 'issue_denomination_s' } %>
+          <%= render partial: 'advanced/facet_as_select', locals: { facet_field_name: 'issue_metal_s' } %>
+          <%= render partial: 'advanced/facet_as_select', locals: { facet_field_name: 'issue_city_s' } %>
+          <%= render partial: 'advanced/facet_as_select', locals: { facet_field_name: 'issue_state_s' } %>
+          <%= render partial: 'advanced/facet_as_select', locals: { facet_field_name: 'issue_region_s' } %>
         </div>
     </div>
     <div class="col-md-7">
     <%= render Blacklight::HiddenSearchStateComponent.new(params: search_state.params_for_search(advanced_search_context.except(:model, :rpp, :start)).except(:f_inclusive)) %>
         <div>
-          <%= render partial: 'facet_as_select', locals: { facet_field_name: 'issue_ruler_s' } %>
-          <%= render partial: 'facet_as_select', locals: { facet_field_name: 'issue_artists_s' } %>
-          <%= render partial: 'facet_as_select', locals: { facet_field_name: 'find_place_s' } %>
-          <%= render partial: 'facet_as_select', locals: { facet_field_name: 'donor_s' } %>
+          <%= render partial: 'advanced/facet_as_select', locals: { facet_field_name: 'issue_ruler_s' } %>
+          <%= render partial: 'advanced/facet_as_select', locals: { facet_field_name: 'issue_artists_s' } %>
+          <%= render partial: 'advanced/facet_as_select', locals: { facet_field_name: 'find_place_s' } %>
+          <%= render partial: 'advanced/facet_as_select', locals: { facet_field_name: 'donor_s' } %>
         </div>
         <%# Publication Date is a range input %>
         <% pub_date = facet_configuration_for_field('pub_date_start_sort') %>
@@ -57,6 +57,6 @@
   </div>
 
   <div class="search-submit-buttons clearfix col-sm-7">
-    <%= render 'advanced_search_submit_btns' %>
+    <%= render 'advanced/advanced_search_submit_btns' %>
   </div>
 <% end %>
