+++ app/views/advanced/numismatics.html.erb
@@ -14,4 +14,4 @@
   contact Alan Stahl, Curator of Numismatics (<a href="mailto:astahl@princeton.edu">astahl@princeton.edu</a>)
   </p>
 </div>
-<%= render 'numismatics_advanced_search_form' %>
+<%= render 'advanced/numismatics_advanced_search_form' %>
